added kepl wallet button
